+++ src/pages/experience/Awards.tsx
@@ -2,13 +2,13 @@ import { useState } from "react";
 
 import { ContentItem } from "../components/content item/ContentItem";
 import { Modal } from "../components/Modal";
-import { volunteerData } from "../../data/volunteer";
+import { awardsData } from "../../data/awards";
 
-export const Volunteer = () => {
-  const [image, setImage] = useState(null);
-  const [caption, setCaption] = useState("");
+export const Awards = () => {
+  const [image, setImage] = useState<string | undefined>(undefined);
+  const [caption, setCaption] = useState<string>("");
 
-  const openModal = (image, caption) => {
+  const openModal = (image: string, caption: string) => {
     setImage(image);
     setCaption(caption);
   };
@@ -31,10 +31,10 @@ export const Volunteer = () => {
 
   return (
     <div className="content-container experience-container">
-      <h1 className="fade left">Volunteer Work</h1>
+      <h1 className="fade left">Awards and Achievments</h1>
       <Modal image={image} caption={caption} setImage={setImage} />
-      
-      {volunteerData.map((project) => (
+
+      {awardsData.map((project) => (
         <ContentItem
           key={project.heading}
           imgOnLeft={imgOnLeft()}
@@ -42,6 +42,7 @@ export const Volunteer = () => {
           subheading={project.subheading}
           dates={project.dates}
           description={project.description}
+          teamSize={project.teamSize}
           githubLink={project.links?.github}
           linkedInLink={project.links?.linkedIn}
           projectLink={project.links?.project}
