+++ src/pages/experience/Experience.tsx
@@ -23,7 +23,7 @@ export const Experience = () => {
     }
 
     window.addEventListener("resize", handleResize);
-    return (_) => {
+    return () => {
       window.removeEventListener("resize", handleResize);
     };
   });
