+++ src/pages/experience/Projects.tsx
@@ -2,13 +2,15 @@ import { useState } from "react";
 
 import { ContentItem } from "../components/content item/ContentItem";
 import { Modal } from "../components/Modal";
-import { awardsData } from "../../data/awards";
+import { projectData } from "../../data/projects";
+import { Sort } from "../components/sort filter/Sort";
 
-export const Awards = () => {
-  const [image, setImage] = useState(null);
-  const [caption, setCaption] = useState("");
+export const Projects = () => {
+  const [image, setImage] = useState<string | undefined>(undefined);
+  const [caption, setCaption] = useState<string>("");
+  const [sortedProjects, setSortedProjects] = useState(projectData);
 
-  const openModal = (image, caption) => {
+  const openModal = (image: string, caption: string) => {
     setImage(image);
     setCaption(caption);
   };
@@ -28,13 +30,14 @@ export const Awards = () => {
     return onLeft;
   };
 
-
   return (
     <div className="content-container experience-container">
-      <h1 className="fade left">Awards and Achievments</h1>
+      <h1 className="fade left">Projects</h1>
       <Modal image={image} caption={caption} setImage={setImage} />
 
-      {awardsData.map((project) => (
+      <Sort projectData={projectData} setSortedProjects={setSortedProjects} />
+
+      {sortedProjects.map((project) => (
         <ContentItem
           key={project.heading}
           imgOnLeft={imgOnLeft()}
@@ -42,7 +45,6 @@ export const Awards = () => {
           subheading={project.subheading}
           dates={project.dates}
           description={project.description}
-          teamSize={project.teamSize}
           githubLink={project.links?.github}
           linkedInLink={project.links?.linkedIn}
           projectLink={project.links?.project}
